Crate seax_schemeUnstable

Library for compiling Scheme programs to Seax SVM bytecode.

This Scheme implementation is based on the Scheme programming language described in the Revised6 Report on Scheme (R6RS). Any cases in which Seax Scheme differs from R6RS Scheme are documented in the RustDoc for the Seax Scheme library and in the Seax Scheme language manual. Do note, however, that some inconsistencies between Seax Scheme and R6RS may be the result of unimplemented features in Seax Scheme.

Modules

ast
Contains the Scheme abstract syntax tree (AST).

parser
Contains the Scheme parser.

Structs

ForkTable
An associative map data structure for representing scopes.

Functions

compile
Compile a Scheme program into a list of SVM cells (a control stack)