Module seax_scheme::astUnstable

Contains the Scheme abstract syntax tree (AST).

The AST stores the semantic structure of a parsed Scheme program, and is responsible for compiling those programs to SVM bytecode instructions, performing semantic analysis (as necessary), and (eventually) for optimizing programs.

Structs

BoolNode
AST node for a boolean constant

CharNode
AST node for a character constant

FloatNode
AST node for a floating-point constant

IntNode
AST node for an integer constant

ListNode
AST node for a list literal

NameNode
AST node for an identifier

RootNode
AST node for the root of a program's AST

SExprNode
AST node for an S-expression.

StringNode
AST node for a string constant

UIntNode
AST node for an unsigned integer constant

Enums

ExprNode
Expression

NumNode

Traits

ASTNode
Trait for AST nodes.

Type Definitions

CompileResult
A CompileResult is either Ok(SVMCell) or Err(&str)

SymTable
The symbol table for bound names is represented as a ForkTable mapping &str (names) to (uint,uint) tuples, representing the location in the $e stack storing the value bound to that name.