Trait seax_scheme::ast::ScopeStable

pub trait Scope<K> where K: Eq + Hash {
    fn bind(&mut self, name: K, lvl: usize) -> (usize, usize);
    fn lookup(&self, name: &K) -> Option<(usize, usize)>;
}

Trait for a symbol table

Required Methods

fn bind(&mut self, name: K, lvl: usize) -> (usize, usize)

Bind a name to a scope.

Returnsthe indices for that name in the SVM environment.

fn lookup(&self, name: &K) -> Option<(usize, usize)>

Look up a name against a scope.

Returns the indices for that name in the SVM environment, or None if that name is unbound.