Module seax_scheme::parser

Contains the Scheme parser.

This parser is based on the Scheme grammar given in the Revised6 Report on Scheme (R6RS). Any deviations from the R6RS standard, especially those with an impact on the valid programs accepted by the parser, will be noted in the parser's RustDoc.

Functions

bool_const
Parses boolean constants.

character
Recognizes R6RS character constants.

expr
Parses Scheme expressions.

float_const
Parser for floating-point constants.

line_comment
Parses a R6RS single-line comment

name
Parser for valid R6RS identifiers.

number
Parses a floating-point, signed integer, or unsigned integer constant.

sint_const
Parser for signed integer constants.

string_const

uint_const
Parser for unsigned integer constants.