Function seax_scheme::parser::nameStable

pub fn name(input: State<&str>) -> ParseResult<NameNode, &str>

Parser for valid R6RS identifiers.

An identifier may begin with an alphabetic character or one of the following special characters !, $, &, :, ^, <, >, _,~,\, or ?. Subsequent characters may also include numbers or the special characters +, -, ., and @.

Essentially, this parser recognizes the regular expression [a-zA-Z!\$%:\^<>_~\\\?][a-zA-Z0-9!\$%:\^<>_~\\\?\+\-\.@]*.

For more information, consult the R6RS.